Re-organize in new dir structure
