--- sw/airborne/modules/computer_vision/OpticFlow/OpticFlowHover.xml
@@ -1,44 +1,0 @@
-<!DOCTYPE module SYSTEM "../module.dtd">
-
-<module name="OpticFlow">
-  <doc>
-    <description>Video ARDone 2</description>
-  </doc>
-
-  <header>
-    <file name="opticflow_module.h"/>
-  </header>
-
-  <init fun="opticflow_module_init()"/>
-
-  <periodic fun="opticflow_module_run()" freq="60" start="opticflow_module_start()" stop="opticflow_module_stop()" autorun="TRUE"/>
-  <makefile>
-    <define name="ARDRONE_VIDEO_PORT" value="2002" />
-    <define name="USE_ARDRONE_VIDEO" />
-      <raw>
-
-include $(PAPARAZZI_HOME)/sw/ext/ardrone2_vision/Makefile.paths
-          
-VISION_MODULE_FOLDER = $(DIR_MODULES)/OpticFlow
-      	     	
-$(TARGET).CFLAGS += -I$(DIR_MODULES) -I$(DIR_CV) -I$(DIR_LIB) -pthread -D__USE_GNU
-$(TARGET).CXXFLAGS += -I$(PAPARAZZI_HOME)/sw/include/ -I$(PAPARAZZI_SRC)/sw/airborne -I$(PAPARAZZI_SRC)/conf/autopilot -I$(PAPARAZZI_SRC)/sw/airborne/arch/$($(TARGET).ARCHDIR) -I$(VARINCLUDE) -I$(ACINCLUDE) -I$(PAPARAZZI_SRC)/sw/airborne/modules/
-$(TARGET).CXXFLAGS += -I$(DIR_MODULES) -I$(DIR_CV) -I$(DIR_LIB) -pthread -D__USE_GNU
-
-$(TARGET).srcs += $(VISION_MODULE_FOLDER)/opticflow_module.c
-$(TARGET).srcs += $(VISION_MODULE_FOLDER)/opticflow_code.c
-$(TARGET).srcs += $(VISION_MODULE_FOLDER)/hover_stabilization.c
-$(TARGET).srcs += $(DIR_CV)/opticflow/optic_flow_ardrone.c
-$(TARGET).srcs += $(DIR_CV)/opticflow/fast9/fastRosten.c
-$(TARGET).srcs += $(DIR_CV)/encoding/jpeg.c
-$(TARGET).srcs += $(DIR_CV)/encoding/rtp.c
-$(TARGET).srcs += $(DIR_CV)/trig.c
-$(TARGET).srcs += $(DIR_LIB)/udp/socket.c
-$(TARGET).srcs += $(DIR_LIB)/v4l/video.c
-$(TARGET).CFLAGS += -I$(DIR_MODULES) -I$(DIR_CV) -I$(DIR_LIB) -pthread
-$(TARGET).LDFLAGS += -pthread -lrt -static
-
-    </raw>
-  </makefile>
-</module>
-
