+++ sw/airborne/modules/computer_vision/opticflow/hover_stabilization.h
